Refactored domain package to contains business rules
