+++ pythonpro/payments/views.py
@@ -9,7 +9,7 @@ from django.urls import reverse
 from django.utils.timezone import now
 from django.views.decorators.csrf import csrf_exempt
 
-from pythonpro import facade
+from pythonpro.domain import user_facade
 from pythonpro.mailchimp.facade import tag_as
 from pythonpro.payments import facade as payment_facade
 from pythonpro.payments.facade import PYTOOLS_PRICE, PYTOOLS_PROMOTION_PRICE, PagarmeNotPaidTransaction
@@ -40,11 +40,11 @@ def pytools_capture(request):
         if user.is_authenticated:
             _promote_client(user, request)
         else:
-            facade.force_register_client(customer_first_name, customer_email, source)
+            user_facade.force_register_client(customer_first_name, customer_email, source)
         dct = {'redirect_url': reverse('payments:pytools_thanks')}
     elif payment_method == 'boleto':
         if not user.is_authenticated:
-            facade.force_register_lead(customer_first_name, customer_email, source)
+            user_facade.force_register_lead(customer_first_name, customer_email, source)
         path = reverse('payments:pytools_boleto')
         qs = urlencode(_extract_boleto_params(pagarme_resp))
         dct = {'redirect_url': f'{path}?{qs}'}
@@ -75,7 +75,7 @@ def _promote_client(user, request):
             'ty_url': request.build_absolute_uri(reverse('payments:pytools_thanks'))
         }
     )
-    facade.promote_client(user, msg)
+    user_facade.promote_client(user, msg)
 
 
 def pytools_thanks(request):
@@ -141,7 +141,7 @@ def pagarme_notification(request, user_id: int):
     except PagarmeNotPaidTransaction:
         pass  # No problem, we need to handle only paid transactions
     else:
-        user = facade.find_user_by_id(user_id)
+        user = user_facade.find_user_by_id(user_id)
         _promote_client(user, request)
     return HttpResponse('')
 
@@ -156,6 +156,6 @@ def pagarme_anonymous_notification(request):
     except PagarmeNotPaidTransaction:
         pass  # No problem, we need to handle only paid transactions
     else:
-        user = facade.find_user_by_email(transaction['customer']['email'])
+        user = user_facade.find_user_by_email(transaction['customer']['email'])
         _promote_client(user, request)
     return HttpResponse('')
